BugFix: process_neighbours skip find_neighbours for init_point NOIE

+++ include/dbscan.h
@@ -11,7 +11,7 @@ typedef struct dataset_t {
 } dataset_t;
 
 typedef struct neighbors_t {
-   int neighbor_count;
+   unsigned long long neighbor_count;
    // int neighbor[ OBSERVATIONS ];
    int *neighbor;
 } neighbors_t;
@@ -54,10 +54,10 @@ unsigned long long dst_cycles;
 
 unsigned long TOTAL_OBSERVATIONS;
 
-double distance( int i, int j );
+double distance( unsigned long long i, unsigned long long j );
 
 // neighbors_t *find_neighbors( int observation );
-neighbors_t *find_neighbors( unsigned long observation );
+neighbors_t *find_neighbors( unsigned long long observation );
 
 void free_neighbors( neighbors_t *neighbors );
